feat(api): restructure API for task and file management
- Remove old api.py file and replace it with a modular structure
- Introduce run_api.py as the new entry point for the FastAPI application
- Create separate modules for task and file management under app/apis
- Implement task and file models for better data handling
- Add routes for tasks and workspace operations with appropriate error handling
- Enhance WebSocket support for real-time updates on task and file changes
- Integrate file monitoring services for directory and file changes
