+++ app/apis/__init__.py
@@ -1,0 +1,3 @@
+from app.apis.routes import router
+
+__all__ = ["router"]
