+++ app/apis/routes/config.py
@@ -1,0 +1,63 @@
+import tomllib
+from pathlib import Path
+
+from fastapi import APIRouter, Body, HTTPException
+
+router = APIRouter(prefix="/config", tags=["config"])
+
+
+@router.get("/status")
+async def check_config_status():
+    config_path = Path(__file__).parent.parent.parent / "config" / "config.toml"
+    example_config_path = (
+        Path(__file__).parent.parent.parent / "config" / "config.example.toml"
+    )
+
+    if config_path.exists():
+        return {"status": "exists"}
+    elif example_config_path.exists():
+        try:
+            with open(example_config_path, "rb") as f:
+                example_config = tomllib.load(f)
+            return {"status": "missing", "example_config": example_config}
+        except Exception as e:
+            return {"status": "error", "message": str(e)}
+    else:
+        return {"status": "no_example"}
+
+
+@router.post("/save")
+async def save_config(config_data: dict = Body(...)):
+    try:
+        config_dir = Path(__file__).parent.parent.parent / "config"
+        config_dir.mkdir(exist_ok=True)
+
+        config_path = config_dir / "config.toml"
+
+        toml_content = ""
+
+        if "llm" in config_data:
+            toml_content += "# Global LLM configuration\n[llm]\n"
+            llm_config = config_data["llm"]
+            for key, value in llm_config.items():
+                if key != "vision":
+                    if isinstance(value, str):
+                        toml_content += f'{key} = "{value}"\n'
+                    else:
+                        toml_content += f"{key} = {value}\n"
+
+        if "server" in config_data:
+            toml_content += "\n# Server configuration\n[server]\n"
+            server_config = config_data["server"]
+            for key, value in server_config.items():
+                if isinstance(value, str):
+                    toml_content += f'{key} = "{value}"\n'
+                else:
+                    toml_content += f"{key} = {value}\n"
+
+        with open(config_path, "w", encoding="utf-8") as f:
+            f.write(toml_content)
+
+        return {"status": "success"}
+    except Exception as e:
+        return {"status": "error", "message": str(e)}
