+++ run_api.py
@@ -1,0 +1,66 @@
+import os
+import threading
+import tomllib
+import webbrowser
+from functools import partial
+from pathlib import Path
+
+from fastapi import FastAPI, Request
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.responses import JSONResponse
+
+from app.apis import router
+
+app = FastAPI()
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+app.include_router(router)
+
+
+@app.exception_handler(Exception)
+async def generic_exception_handler(request: Request, exc: Exception):
+    return JSONResponse(
+        status_code=500, content={"message": f"Server error: {str(exc)}"}
+    )
+
+
+def open_local_browser(config):
+    webbrowser.open_new_tab(
+        f"http://{config.get('host', 'localhost')}:{config.get('port', 5172)}"
+    )
+
+
+def load_config():
+    try:
+        config_path = Path(__file__).parent / "config" / "config.toml"
+
+        if not config_path.exists():
+            return {"host": "localhost", "port": 5172}
+
+        with open(config_path, "rb") as f:
+            config = tomllib.load(f)
+
+        return {"host": config["server"]["host"], "port": config["server"]["port"]}
+    except FileNotFoundError:
+        return {"host": "localhost", "port": 5172}
+    except KeyError as e:
+        print(
+            f"The configuration file is missing necessary fields: {str(e)}, use default configuration"
+        )
+        return {"host": "localhost", "port": 5172}
+
+
+if __name__ == "__main__":
+    import uvicorn
+
+    config = load_config()
+    open_with_config = partial(open_local_browser, config)
+    threading.Timer(3, open_with_config).start()
+    uvicorn.run(app, host=config["host"], port=config["port"])
